+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - Show the number of selectable poll answers as a subtitle in poll message content. [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
 - Show unique poll winner. [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
 - Show "Vote ended" subtitle for closed polls.. [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
+- Shows the vote date and time in the Poll result dialog. [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
 
 ### ✅ Added
 
@@ -81,6 +82,7 @@
 - Show "Vote ended" subtitle for closed polls.. [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
 - Show the number of comments on the "View Comments" button. [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
 - Show discard dialog when there are changes in the poll composer and the system back button is "pressed". [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
+- Shows the vote date and time in the Poll result dialog. [#5872](https://github.com/GetStream/stream-chat-android/pull/5872)
 
 ### ✅ Added